Removing problematic libraries from original source. VRC7 and S5B are now disabled.
The original source came bundled with proprietary (yet available) libraries that are
incompatible with the GPL.
As the components are not derived from this project, it's not illegal to distribute
the components independently (according to the GPL).
However, it's questionable at best to compile that code with GPL'd code. I'm taking
a precautionary measure by removing that code.
What's removed:
- A fast fourier transform library - we're not using this
- The ym2149 and ym2413 emulator libraries, which VRC7 and S5B rely on.
  - TODO: Take the time to analyze the libraries and write a new, GPL-compatible
    implementation.