Starting sound generation, IO seeking
Instead of putting sound playback directly into the sound generation class,
I'm creating a SoundSink class so to separate sound playback with sound
generation. This is particularly important from a cross platform perspective,
because we might want to write an accompanying OpenAL or SDL fallback engine, or
the respective engines for Windows and Mac OS X.
Sound exporting is done this way as well.