FtmDocument: Make sure to deprecate channel stuff from FtmDocument and move into SoundGen
FtmDocument should never have to query a sound generator class. What if we wanted to simply
read the document, and not play anything? Nothing in the class except modifying a tracker channel's
effect columns REALLY depends on those variables being there.
However, the document class should only exist to read and write music information. It should never
interact with the application on our behalf. If we want to modify the effect columns in two places,
then we should do so at a higher level. An wrapper at the gui level would be better.