qt-gui: gui progress
FrameView: I'm pretty happy with the way this is turning out. The colors are
hard coded in for the moment, but I'll change that the same time I
implement color styles.

MainWindow: Song information, instruments, pattern inc/dec, etc. Good!

gui: Seting up a list of "loaded documents". We only currently show one
at a time, but we should leave room for being able to support concurrent
documents. For now, we have the activeDocument() and activeDocInfo() functions
for the widgets to use.