Sound: Added new TrackerController class. Also, music playback works!
The TrackController class contains all the playback variables for playing
the current track. It's also meant to be simpler to get the current frame and
row, so not to overcomplicate things with events.

Feeding SoundGen a document and sound sink and calling run() successfully plays
most songs. The original enl_track1.ftm doesn't automate, which means I'm probably
missing something. Resaving it into the newer version plays it successfully.