IO: isReadable() and isWritable(); sound: Tracker update callback
The isReadable/Writable methods are for knowing if we can start reading
or writing to a file before we do so.

The tracker update callback is called when the position of the tracker
changes. The implementer can choose to send an event to the UI, or
just do something synchronously.