gui: PatternView renders text every frame again, shows next frame, FrameView prevents modification during playback
For now, PatternView renders the text directly instead of using QPixmap.
This is because of the more noticable rendering lag when entering a new frame with
songs with longer patterns (64-256 rows).
A new solution will probably involve a separate QPixmap for each row, or for each
group of rows.

PatternView will also show the next frame in darker colors. The previous frame
is TODO.