gui: DocInfo::scrollFrameBy, threads: set running flag before running thread, PatternView: highlight rows
scrollFrameBy: there are multiple instances where it's useful to simply offset the rows
(scroll wheel, up/down keys).

threads running flag: For some reason, pressing the enter key to play the song would sometimes glitch up the
application and crash. It would sometimes let you edit the row when you weren't supposed to.
I assumed it was a problem with isPlaying(), so I enabled the running flag before starting the thread, to ensure
it was set for the main thread.

PatternView: only render the necessary highlight rows