gui: Random deadlock while playing (most likely) fixed.
This is why it's important to lock your mutexes when using conditionals.
The update event might try to notify the conditional before we can wait
on it. This scenario is possible if the update event does nothing and/or
the sendUpdateEvent() thread is interrupted by the OS after the postEvent.
The chance of it happening in one shot is slim, but compounded enough
times and it happens frequently enough to become annoying.
It's more likely to happen in slower songs because more of the
update events take less time to perform.