gui: Playback commands on thread pool; sound/alsa: addressing deadlock issues on shutdown
gui thread pool: Thread pools are better than creating a new thread for each command because
it better ensures commands are done sequentially. It's also a cleaner and less bug prone
implementation.

There was (possibly still is) a bug where the sound sink timer would continue adding times
after the gui shut down. If the gui is unable to respond to the time callback, the thread
deadlocks, meaning the program is unable to shut down.
It has something to do with SoundGen calling SoundSink::setPlaying(false), and the ALSA
sound sink causing the problem.