Adding VRC7 support, including emu2413
One of the reasons VRC7 (and S5B) wasn't included earlier was because of potential
copyright/licensing issues. emu2413.c doesn't explicitly include a copyright license
(be it LGPL, MIT, BSD, public domain, etc.).
The source of the file originates from the MAME project, in which developers may individually
license their contributions.
The license (zlib) can be found: http://mednafen.sourceforge.net/documentation/09x/mednafen.html
I believe Mednafen to be a good authority on the matter, and Google seems to agree with the consensus
regarding the license for emu2413.c