ncurses-ui: Thread pool implemented
A thread pool would be required for any sort of interaction with a keyboard (or mouse).
One thread runs to capture keys (input) and dispatch key events, and
another thread (tracker update) dispatches update events.
These are currently implemented to allow the user to exit out of the program without
resorting to Ctrl+C.
There should preferrably be one unifying thread pool design for both the ncurses and Qt ui.
I like this design better, so this design will probably be put into the Qt version.